# src/generated/java/com/ebs/controller/EmployeesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.ebs.controller;

import com.ebs.model.EmployeeDTO;
import com.ebs.model.EmployeesResultModel;
import com.ebs.model.ErrorDTO;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Optional;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-12-20T18:43:45.486+01:00")

@Api(value = "employees", description = "the employees API")
public interface EmployeesApi {

    Logger log = LoggerFactory.getLogger(EmployeesApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "", nickname = "createEmployee", notes = "Create employee", response = EmployeeDTO.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Employee Created", response = EmployeeDTO.class),
        @ApiResponse(code = 400, message = "Bad Request", response = ErrorDTO.class),
        @ApiResponse(code = 401, message = "Access is denied due to invalid credentials", response = ErrorDTO.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDTO.class) })
    @RequestMapping(value = "/employees",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    default ResponseEntity<EmployeeDTO> createEmployee(@NotNull @ApiParam(value = "Name of the employee", required = true) @Valid @RequestParam(value = "name", required = true) String name,@ApiParam(value = "surName of the employee") @Valid @RequestParam(value = "surName", required = false) String surName,@ApiParam(value = "email of the employee") @Valid @RequestParam(value = "email", required = false) String email,@ApiParam(value = "address of the employee") @Valid @RequestParam(value = "address", required = false) String address,@ApiParam(value = "salary of the employee") @Valid @RequestParam(value = "salary", required = false) Integer salary,@ApiParam(value = "companyId of the employee") @Valid @RequestParam(value = "companyId", required = false) Integer companyId) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{  \"companyId\" : 1,  \"address\" : \"address\",  \"surname\" : \"surname\",  \"name\" : \"name\",  \"id\" : 0,  \"salary\" : 1,  \"email\" : \"email\"}", EmployeeDTO.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default EmployeesApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "", nickname = "deleteEmployee", notes = "Delete employee ", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Employee deleted"),
        @ApiResponse(code = 400, message = "Bad Request", response = ErrorDTO.class),
        @ApiResponse(code = 401, message = "Access is denied due to invalid credentials", response = ErrorDTO.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDTO.class) })
    @RequestMapping(value = "/employees/:id",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deleteEmployee(@NotNull @ApiParam(value = "Id of the employee", required = true) @Valid @RequestParam(value = "id", required = true) Integer id) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default EmployeesApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "", nickname = "editEmployee", notes = "Edit employee", response = EmployeeDTO.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Employee Created", response = EmployeeDTO.class),
        @ApiResponse(code = 400, message = "Bad Request", response = ErrorDTO.class),
        @ApiResponse(code = 401, message = "Access is denied due to invalid credentials", response = ErrorDTO.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDTO.class) })
    @RequestMapping(value = "/employees",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    default ResponseEntity<EmployeeDTO> editEmployee(@NotNull @ApiParam(value = "Id of the employee", required = true) @Valid @RequestParam(value = "id", required = true) Integer id,@ApiParam(value = "surName of the employee") @Valid @RequestParam(value = "surName", required = false) String surName,@ApiParam(value = "email of the employee") @Valid @RequestParam(value = "email", required = false) String email,@ApiParam(value = "address of the employee") @Valid @RequestParam(value = "address", required = false) String address,@ApiParam(value = "salary of the employee") @Valid @RequestParam(value = "salary", required = false) Integer salary,@ApiParam(value = "companyId of the employee") @Valid @RequestParam(value = "companyId", required = false) Integer companyId) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{  \"companyId\" : 1,  \"address\" : \"address\",  \"surname\" : \"surname\",  \"name\" : \"name\",  \"id\" : 0,  \"salary\" : 1,  \"email\" : \"email\"}", EmployeeDTO.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default EmployeesApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "", nickname = "getEmployee", notes = "Get", response = EmployeesResultModel.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Get all employees", response = EmployeesResultModel.class),
        @ApiResponse(code = 400, message = "Bad Request", response = ErrorDTO.class),
        @ApiResponse(code = 401, message = "Access is denied due to invalid credentials", response = ErrorDTO.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDTO.class) })
    @RequestMapping(value = "/employees",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<EmployeesResultModel> getEmployee() {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{  \"EmployeeDTO\" : [ {    \"companyId\" : 1,    \"address\" : \"address\",    \"surname\" : \"surname\",    \"name\" : \"name\",    \"id\" : 0,    \"salary\" : 1,    \"email\" : \"email\"  }, {    \"companyId\" : 1,    \"address\" : \"address\",    \"surname\" : \"surname\",    \"name\" : \"name\",    \"id\" : 0,    \"salary\" : 1,    \"email\" : \"email\"  } ]}", EmployeesResultModel.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default EmployeesApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "", nickname = "getEmployeeForId", notes = "Get", response = EmployeesResultModel.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "GetEmployee for id", response = EmployeesResultModel.class),
        @ApiResponse(code = 400, message = "Bad Request", response = ErrorDTO.class),
        @ApiResponse(code = 401, message = "Access is denied due to invalid credentials", response = ErrorDTO.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDTO.class) })
    @RequestMapping(value = "/employees/:id",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<EmployeesResultModel> getEmployeeForId(@NotNull @ApiParam(value = "Id of the employee", required = true) @Valid @RequestParam(value = "id", required = true) Integer id) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{  \"EmployeeDTO\" : [ {    \"companyId\" : 1,    \"address\" : \"address\",    \"surname\" : \"surname\",    \"name\" : \"name\",    \"id\" : 0,    \"salary\" : 1,    \"email\" : \"email\"  }, {    \"companyId\" : 1,    \"address\" : \"address\",    \"surname\" : \"surname\",    \"name\" : \"name\",    \"id\" : 0,    \"salary\" : 1,    \"email\" : \"email\"  } ]}", EmployeesResultModel.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default EmployeesApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}